tui: separate minimal and extra themes into modular draw/ files
